rename files and separate concerns
